Add day 1 solve retrospectively
I did this one using macros in nvim and bc. Unfortunately, I overrode
solve1.txt when solving day 2, so I didn't save the original results,
and I also didn't commit the macros which were the cooler part of this
challenge. But it was fairly straightforward, and for solve 2 you can
see more complicated formulas which I piped to bc.